+++ src/php/Gdbots/Schemas/Pbjx/Command/CheckHealth.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace Gdbots\Schemas\Pbjx\Command;
+
+use Gdbots\Pbj\Message;
+
+interface CheckHealth extends Message
+{
+}